Use 32bit int for fields in TIFF structures rather long which can be 32 or 64bit.

+++ test/com/ComModTestAAF/ModuleTests/CAAFTIFFDescriptorTest.cpp
@@ -2,7 +2,7 @@
 // @com This file implements the module test for CAAFTIFFDescriptor
 //=---------------------------------------------------------------------=
 //
-// $Id: CAAFTIFFDescriptorTest.cpp,v 1.27 2004/02/27 14:26:51 stuart_hc Exp $ $Name:  $
+// $Id: CAAFTIFFDescriptorTest.cpp,v 1.28 2004/10/25 10:09:20 stuart_hc Exp $ $Name:  $
 //
 // The contents of this file are subject to the AAF SDK Public
 // Source License Agreement (the "License"); You may not use this file
@@ -40,6 +40,9 @@ using namespace std;
 
 #include "CAAFBuiltinDefs.h"
 
+// Reference for TIFF data structures is Adobe's TIFF revision 6.0
+// http://partners.adobe.com/asn/developer/pdfs/tn/TIFF6.pdf
+
 #define	TIFF_VERSION	42
 
 #define	TIFF_BIGENDIAN		0x4d4d
@@ -49,14 +52,14 @@ typedef	struct
 {
 	unsigned short tiff_byteOrder;		
 	unsigned short tiff_identNumber;	
-	unsigned long  tiff_firstIFD;	/* byte offset to first directory */
+	unsigned int   tiff_firstIFD;	/* byte offset to first directory */
 } TIFFHeader;
 
 typedef	struct {
 	unsigned short tdir_tag;	/* see below */
 	unsigned short tdir_type;	/* data type; see below */
-	unsigned long  tdir_count;	/* number of items; length in spec */
-	unsigned long  tdir_offset;	/* byte offset to field data */
+	unsigned int   tdir_count;	/* number of items; length in spec */
+	unsigned int   tdir_offset;	/* byte offset to field data */
 } TIFFDirEntry;
 /*
  * Tag data type information.
